client: make less noise when unlinking during readdir
Skip, but do not talk about, NULL dentries.

+++ src/client/Client.cc
@@ -832,11 +832,15 @@ void Client::insert_readdir_results(MetaRequest *request, MetaSession *session, 
 	if (pd->first < dname &&
 	    diri->dirfragtree[ceph_str_hash_linux(pd->first.c_str(),
 						  pd->first.length())] == fg) {  // do not remove items in earlier frags
-	  ldout(cct, 15) << "insert_trace  unlink '" << pd->first << "'" << dendl;
 	  Dentry *dn = pd->second;
-	  ++pd;
-	  unlink(dn, true, true);  // keep dir, dentry
-	  dn->lease_mds = -1;
+	  if (dn->inode) {
+	    ldout(cct, 15) << "insert_trace  unlink '" << pd->first << "'" << dendl;
+	    ++pd;
+	    unlink(dn, true, true);  // keep dir, dentry
+	    dn->lease_mds = -1;
+	  } else {
+	    ++pd;
+	  }
 	} else {
 	  ++pd;
 	}
